Fix SharedIssuePanelTest and remove bleak check
This moves the SharedIssuePanelTest test to a non-bleak test. Once the
test is fixed, it keeps failing because of the Bleak test. The test
seems to be failing the bleak check for reasons that are not related to
the issue panel. Even opening and closing a tab fails the test so, while
we debug the bleak problem, making this a non-bleak check.

Test: SharedIssuePanelTest
Bug: 270432776
Change-Id: I72a09aa6bf0d1f907160eb2e6f6e0747c4207972