Use a project system token for finding dependencies with resources
The current entry point has a module as the unit of enquiry, which
cannot be right, but its wrongness should not additionally mean
transitive support of the superset of things one can ask of a module.
Move the Gradle-specific implementation to project-system-gradle
through a token, while preserving the existing behaviour for all
other project systems

This is a part of removing some Gradle-specific module-per-source-set
interfaces from ModuleSystemUtils.

Bug: 147586218
Test: existing
Change-Id: Ia5b07b874f73e88a03116a28b894f3889a480a12