Disable drag-and-drop for unsupported resources
The drag-and-drop functionality in resource explorer is supposed to
trigger the "import drawable" dialog. The problem is this dialog is
shown regardless of the tab selected in resource explorer, causing
some confusion, for example, if the user tries to drag a layout file
when the "Layout" tab is selected.

This change modifies the behavior to only allow drag-and-drop when
tabs corresponding to valid resources, e.g. Drawables and Mip Maps,
are selected.

Fixes: 244289961
Test: N/A. Manually verified by drag-and-drop resources to explorer.
Change-Id: I2daa57c11f686bf22ab9bed879068d5b03e793fe