Hide the floating modes in AttachedToolWindow.
The floating tool windows in Intellij is used for implementing
the floating mode of our attached tool windows in our designers
and inspectors. The Intellij versions do not offer the flexibility
to override the actions. This leads to a bad experience when the
iuser wants to attach the tool windows again.

In this CL: remove the floating tool window support. Add a flag
such that we can turn in on if someone really needs it.
If the flag is off we should ignore floating and detached states
saved from an earlier run.

Fixes: 228094557
Test: N/A
Change-Id: I61b79f4045c08593bd18e052fa0ce9eebe42f480