Use a project system token for finding icons
Elements, such as PsiElements or Modules, when shown in various UI
controls, can have icons associated with them.  The appropriate icon
to show can depend somewhat on the project structure (for example, a
sourceSet module can have an icon specialized to its sourceSet type),
which is project-system specific, so provide an interface to allow UI
clients to request appropriate icons.

Use the Gradle implementation directly from subsystems which already
have a dependency on the Gradle project system, but request the
generic token from android core.

This is a part of removing some Gradle-specific module-per-source-set
interfaces from ModuleSystemUtils.

Bug: 147586218
Test: existing
Change-Id: Ie8222e0516514369d244f9c43e3957bb1be6c50d