Faster `removeAll` for our use of `IdentityHashMap`
`IdentityHashMap.removeAll` always iterates over the receiver map even
when the set to remove is smaller. For our use of `IdentityHashMap`,
the optimization would be ok, because we collect the instances from
the same tree.

This change optimizes the calls to `removeAll` to mitigate the scenario
when `remainders` is large and we repeatedly iterate over it at each
node when calling `removeAll`. The optimized version should be no
worse than the original one even if it doesn't fix the reported freezes.

Bug: 341293469
Test: n/a
Change-Id: I4645e5d5de55dcc792c8efe250ec607971bfe63d