Force debuggable build & launch action when api level < 29
Before this change, if the user initiates a startup task with a task
that prefers a profileable build, but the device does not support
profileable builds (api < 29), the user is prompted to continue with
debuggable. With this change, using the feature level of the device,
if the device's api < 29, it falls back to debuggable build to prevent
such prompting.

Fixes: 343611676
Test: manual
Change-Id: I776c67bf3b3a7c10d58d417d27dc298efc565b7c