Use PsiAwareTextEditorProvider instead of TextEditorProvider
This change causes us to use PsiAwareTextEditorProvider instead
of TextEditorProvider when SourceCodeEditorProvider creates a
TextEditor. In the IntelliJ platform, there are a suite of
TextEditorInitializer classes that initialize a newly created
TextEditor. This includes things like:

* Rendering doc comments
* Setting up folding
* Rendering declarative inlay hints
* Rendering parameter hints
* and much more!

Unfortunately, these TextEditorInitializer classes are invoked
only for the PsiAwareTextEditor. This happens in the
createEditorBuilder method of PsiAwareTextEditorProvider.

By using this provider instead, we can ensure that each
TextEditorInitializer gets a chance to run.

Fixes: 339313333
Test: Manually tested
Change-Id: Ibff4e480b256d01c536b0b24e7b225793d8f84b4