Do not touch the file system in Sync Status Notification Provider
Rather than calling a common utility function which expects an ioFile,
and which then performs a filesystem operation to see whether that
file exists, from the EDT, use the fact that we are being called from
a context where we hold the read lock, we know that the virtual file
is being edited in a visible editor, and the virtual file's validity
has already been checked within the read action.  This means that we
can elide the filesystem check, and merely check the file names.

(Technically this simplification is invalid if the user has somehow
opened an editor on a *directory* named by one of the special Gradle
file names.  I suspect that this would be the least of that user's
problems.)

Fixes: 343281453
Test: updated
Change-Id: I60c4d82089555feba4a1e45576292806e357680c