Make inner classes static
This breaks the implicit relationship between
ResourceNotificationManager and these observer classes, and calls out
the required interactions explicitly (in the form of method references).

There is no functional change in behavior, with the exception of
`ProjectPsiTreeObserver.ignoreChildrenChanged`. Previously that flag was
stored on ResourceNotificationManager, so in the case where the observer
is removed and then recreated the old value of the flag would be
persisted; now it always resets to false when the observer is
constructed. Given how the observer operates, this default value should
not make a difference.

Bug: n/a
Test: n/a
Change-Id: I16ead00b45cb75d2e28ea690856c1ed7b3d72e16