Implement conservative copy of transport files
Currently, each time the transport is started (or an agent is started)
each file is always copied to the device, even if the device already
contains an exact copy.

This change implements conservative copy, which before pushing a file
calculates its hash and compares it with the hash of the file on the
device. If the hash is the same there is no need to push the file.

Each time a file is pushed to the device, an hash file is created.
The hash file contains the hash of file that was just pushed. Before
pushing a new file, we get the hash from the hash file and compare it
with the newly computed hash for the file we want to push.

An alternative, and slightly more efficient, approach would have been to
store the hash of the file directly in the name of the file. This
approach is more complicated because throughout the codebase we make
assumptions on what we expect the file names to be.

After this change the average initialization time for the transport
(measured from TransportFileManager#copyFilesToDevice) is 982 ms.
Before this change the average initialization time is 3026 ms.

Measurements for Layout Inspector:
Fist time time to connect goes to 4237 ms from 6695 ms.
Non first time time to connect goes to 1160 ms from 1434 ms.

Test: NA
Bug: NA
Change-Id: Ia18c8e2b15530a39cf0af3b3f72674f310bae0f2