[AUA] Set the projectBuildModel context agpVersion early
As the Dsl model is required to evaluate Dsl builds with different
implied AGP versions, we need to provide the agp version to the model
in order to be able to accurately compute what the build is.  This is
necessarily part of the initialization of any
GradleBuildModelRefactoringProcessor, so adjust the abstract base
class to be instantiated with the current AGP Version.

In particular this means that we will no longer spuriously create
`packagingOptions` blocks for AGP Upgrades involving a current AGP
with version beyond 8.0.0-beta02.  However, this does not address the
issue where an upgrade straddling the 8.0.0 release will have some
elements created using the old pre-8.0 operators, though this is
largely a cosmetic issue.

Fixes: 343922414
Test: added
Change-Id: I2184e5124b5d4e575c5a881c9c48914db5a6960e