[Dsl] Make the model for viewBinding.enabled version-specific
This property was removed in AGP 7.0, but Android Studio needs to
support projects building with older versions of AGP.  This means that
client code will legitimately want to refer to this property, but also
might not be in the best position to know about its replacement.

Use the model's context (specifically its AGP version) to decide what
to do.  If the context's AGP version is null, or known to be older
than the buildFeatures block, simply model the property directly.  If
the AGP version is known to be newer than when the property was
removed, redirect to the new property.

In between, model the viewBinding.enabled property if it already
exists on file (or was created by something else); however, if it does
not exist, model the new property.

Bug: 142099752
Test: added
Change-Id: I37fd340256e75cbadef01557d82c5c58c7048b7a