Reduce lock scope and factor out common observers
This class makes use of multiple "Observers", as well as maps that are
operating similarly to a MultiMap. They currently are all modified
within one giant lock, whose purpose is to (1) ensure consistency in
listener collections, and (2) initialize event registration on demand as
it's needed and remove it when all relevant listeners have been removed.

This change moves common logic from the maps and observers into super
classes, and rearranges the locks to be much more granular and only
include modifications the the listener collections. Event
registration is generally *not* done under the locks any more. (This
change also modifies the observers to be single-use; so if an observer
has its last listener removed it will become obsolete and a new one will
be created in the future if a new listener is added. Being single-use is
what allows event registration to move outside the lock, since we don't
need to worry about a new listener coming in.)

Bug: 343653602
Test: ResourceNotificationManagerTest
Change-Id: Ic5091180278fd42bf97d979977091e76abf78540