Move other resource initialization outside locks
There are some pieces of per-facet initialization that are done whenever
the first listener for a given facet is register. Doing these inside a
lock has the potential for deadlock (and in fact that happened in the
tagged bug).

There's no need for these to be done inside the lock; the lock is just
ensuring consistency of listener collections. And in fact it's safe to
do this initialization on every listener add, since it's only just
fetching a couple services; after the first time, the service fetching
won't actually run any logic other than returning the existing object.

Bug: 343653602
Test: ResourceNotificationManagerTest
Change-Id: Id0f654afaf331d3e4f3a287dec9252b00514d1f5