Properly handle rememberComposableLambda
Also adds code to handle ComposableLambdaN, even though I tried and
could not force Compose to generate these calls even with
functions/nested Composables with absurd numbers of arguments.

Also makes the invalidation action for non-Compose changes a SAVE AND
LOAD as it is in the non-bytecode analysis path, to make debugging
differences between the paths easier.

Bug: 343491010
Bug: 346068999
Test: Manual, unfortunately; could not figure out how to get Compose to
generate either of these calls
Change-Id: Iaec18df8d9f239d6b6bd3f08c8d28dba0ce50b12

Change-Id: I7a7366da602b45f6ac14d1ed728436983edb57ea