Use thread safe Set in DesignerTypeRegistrar
There are multiple threads writing to the set of registrars.
In this UI test failure the Layout file type was omitted because
the Navigation file type was added at the same time.

Therefore the layout file in the test was not identified as a
layout and the palette failed to initialize for layouts.
i.e. none of the categories are present.

Fixes: 345579626
Test: ComponentSelectionSplitViewTest
Change-Id: I911d3846eae209ca6221c5bfe320e1fcd7cff8e0