Provide an inputEvent in FindSelectedLibVersionIntegrationTest
With IntelliJ change f2c64512cee233ae8f4c72515738df1d161c4d9d
the KeyEvent used to trigger this action in tests can't be
instantiated since we aren't passing in a valid InputEvent or
Component associated with the Action. To get around this, a
stub InputEvent is created to pass along with the action.

Change-Id: Iec104a0f2ed6356b6129f9d0516b8250792293cd