Prevent "Start anyway" text when cannot start task
When there is an error with task configuration selection,
disabling the "Start profiler task" button, an error
icon + message is shown. However, if the process selected
is debuggable and the task selected is profileable preferred,
it also changes the text to "Start anyway". This change of
text is tied to the profileable-preferred warning, but
should be superceeded by the error. Thus, the text
should stay "Start profiler task".

Fixes: 344986934
Test: manual
Change-Id: I6819a6edcc23ba9f32ab2facedf8add848463bfb