Type signature changes for test to sourceProvider maps
Define HostTest and DeviceTest interfaces for use in type signatures,
distinct from the implementation enum classes.

Propagate the new HostTest and DeviceTest interfaces through the
system, so that it should no longer be possible to accidentally
attempt to look up a host test source provider from a device test map,
or the reverse.

Fixes: 343189827
Test: existing (no change in functionality)
Change-Id: I431077e1c697098c357b181cdf3b7e53a3c1e7e0