Make tooltip hitbox include text for non-collapsed notifications
As of now, this only covers the case of the profileable preferred warning.
It will have text + an icon if not collapsed, as well as a tooltip.
Before this change, one could only mouse over the icon for the tooltip.
Now, they can mouseover both the icon and the tooltip.

Fixes: 348022955
Test: manual
Change-Id: Ie08aa3d24e4b4d58580eb46afbcb05bde980f2f9