Refactoring to BUILD_OUTPUT_WINDOW_STATS gathering
Collect the generated events in View listener instead of the
wrappers, parse metrics values from the events received by the view.
This will help to report all that is shown, not only what is generated
from the parsers (there are other ways to generate error messages and
not all the parsers are wrapped theoretically).
Additionally, this removes the necessity to keep state and clear it at
the right times.

Share this code with metrics gathered for sync failure details.

Bug: 326938231
Test: Adjusted related tests
Change-Id: I02ba52ee38133cc16e24c4fd2efda49e384ee6ce