Fix race condition in NavStatusCache.
The cached data in NavStatusCache is invalidated when
NavigationResourcesModificationListener raises an event. This is
triggered when a navigation resource file changes. However,
NavStatusCache computes its information based on
StudioResourceRepositoryManager's view of resources, which are updated
on a background thread _after_ the file change. This can result in
NavStatusCache reading and caching bad data, if it fills the cache after
the PSI file changes, but before StudioResourceRepositoryManager
finishes scanning and updating the file.

To fix this, defer NavigationResourcesModificationListener's event until
StudioResourceRepositoryManager is done with its processing. This will
ensure that NavStatusCache has fully up-to-date information available
when it caches its information.

Fixes: 335882184
Test: bazel test --runs_per_test=500
  //tools/adt/idea/nav/safeargs/tests/...:all
Change-Id: I2aed049ca8376c779060b119a852a14b77218962