Fix TemplateDiffTest to use [plugins]
Previously, plugins were not actually applied as plugins in TemplateTest
and then TemplateDiffTest. This change adds an empty plugins {} block in
build.gradle so that DependenciesHelper will apply the plugin with a
version and put that in the [plugins] block of libs.versions.toml.
settings.gradle is also changed so the plugin artifacts can be found
during offline tests.

Bug: 339781996
Test: TemplateDiffTest
Change-Id: I3292ec166a09340aa30de49e351f8cfb1ce9d0b4