Remove redundant AndroidFacet utilities
They are redundant given the existence of equivalent ModuleSystemUtil
methods, which are themselves in the process of being tidied up to
LinkedAndroidModuleGroupUtils in the Gradle project system (since
these linked modules are a Gradle-specific construct).  Reducing the
number of ways these concepts can be spelt by one will help in the
tidying process.

Bug: 147586218
Test: existing (no change in functionality)
Change-Id: If0ba38a72ba277076f384e6a91cf09553519e861