Move LinkedAndroidModuleGroup to the project system
Now that AndroidFacet helper methods have been removed, it no longer
needs to live in android-common.  Sadly the generic project system is
still generally-accessible, if a little bit more hidden away; until
all the users of getMainModule() and friends are addressed we cannot
put this into the Gradle project system where it belongs.

Bug: 147586218
Test: existing (no change in functionality)
Change-Id: I8ab401159c35fef5edd0730d5324c1648b815940