Read resource directories from the facet via a project system token
The facet has the up to date source provider information, which
captures (for Gradle projects) source providers for all source sets --
so to get the right subset of source providers we need to know which
source set we are actually looking for.  This is implicit in the
ResourceFolderManager's `module`, but making that implicit knowledge
explicit should be a responsibility of the project system, not the
client code.  This change implements a Project System Token to perform
this choice.

Bug: 147586218
Test: existing
Change-Id: I453e2c875564eaa0f2de4342a802b3271bfa1210