Remove reliance on arePropertiesSet in WearPairingManager
The arePropertiesSet method returns true after a property has been
fetched and the properties cache has been populated.

After the change from ddmlib to adblib, properties are not always
fetched when the device is booted. Furthermore, while a property is
being fetched, arePropertiesSet can switch to false temporarily while
the cache is being invalidated and re-populated.

In WearPairingManager, we call the arePropertiesSet method to check if
a device has booted. The fact that the method can sometimes return false
after having returned true before or that it can return false just
because no property has been fetched yet, means it's not the best
candidate for checking if a device has booted.

This CL changes WearPairingManager to rely on IDevice.isOnline instead.

Fixes: 350735240
Test: added regression test to EndToEndIntegrationTest
Change-Id: If06e4c20df01b7e8c65576fffbac6926ab339ec7