Allow K2 ComposeDeviceSpecInjectionContributor to call AA on write-action
ComposeDeviceSpecInjectionContributor is used as a part of code-format
that can be also a part of template execution. Since the user can
update PSIs for the template, the code-format can happen right after the
PSI updates. In that case, we cannot run analysis APIs (AA) for the
updated PSIs in advance on a background thread, which means we cannot
avoid running analysis APIs on the UI-thread. In addition, the
code-format runs on a write-action. This commit allows K2
ComposeDeviceSpecInjectionContributor to call AA on write-action, which
is unavoidable.

Bug: 345812375
Test: PreviewPickerTests
Change-Id: Ic0a2005a291b2cf0b25c82a57a6e45038435e6d7