Fix InvalidVirtualFileAccessException during tests execution
With IntelliJ 2024.2 the KLIB manifest file type was introduced as
part of KT-62936. However, a file that was already deleted is
specified on the KlibManifestFileType.isMyFileType while
ensuring that indexes are up-to-date on CodeInsightTestFixtureImpl,
resulting in InvalidVirtualFileAccessException for accessing an invalid
virtual file that no longer exists.

This happens because during starting test execution where
CodeInsightTestFixtureImpl.ensureIndexesUpToDate invokes force update
on FileBaseIndexImpl files from the previous test execution,
keeping those references to a deleted temporary directory.
Those references are kept on myChangedFilesCollector and
myFilesToUpdateCollector and aren't properly cleaned up between executions.

For this reason, this CL ensure that a proper cleanup of those
references happen on the test tear down.

Despite that KlibManifestFileType.isMyFileType implementation should
check if file exists before finding the 'linkdata' children name
(KT-69891), this issue only affects tests execution.

Bug: 352709056
Test: Existing
Change-Id: I0ed84f4b48db2edf1ada99e595f9d65e39e841c7