Fix CommonPreviewRepresentationTest after Intellij 2024.2
Similar to Ia87da45f49f3ab0125fca34c7, the DataContext needs to be of a
known type to be used by CustomizedDataContext. In this case the mocked
type was not of one of DataManager.getDataContext,
CustomizedDataContext, or SimpleDataContext.

This CL replaces the use of the mocked DataContext with an empty
SimpleDataContext.

Fixes: 353201458
Test: CommonPreviewRepresentationTest
Change-Id: If8b1b4930c1600eb716de3afd36c99e2c8b55608