Remove obsolete LifecycleUsage handling
The attribute being checked here was deprecated 6 years ago, and this
was never changed to handle the new attribute; so this logic was already
effectively broken for any users with up to date dependencies.
Furthermore, the replacement attribute was also subsequently deprecated
about 3 years ago and replaced with a more normal listener pattern (ie,
implementing an interface and registering the listener). Given that,
there's no longer a need to check for this annotation.

Bug: 353272351
Test: AndroidJavaHighlightingTest
Change-Id: I015c9822f1512fa6de286e651de2e2f3cba97456