Rewrite ComposeCreatePreviewActionK2 to fix memory leak
The existing ComposeCreatePreviewActionK2 keeps a smart pointer to a
PsiElement because we do not have a ModCommandAction that separates the
availability check of the action from the PSI update phase. Since the
availability check needs analysis APIs (AA), running them together with
the PSI update (write-action) causes wrong analysis results e.g., PSI
can be updated after running AA. The existing
ComposeCreatePreviewActionK2 relies on PsiBasedModCommandAction to
separate AA from the write-action, but it does not provide us with an
API to keep the target PSI element. It has a member property of smart
pointer to the PSI element. Since we do not have an API (like dispose
function) to clean it up, it keeps pointing a PSI element, which causes
a memory leak.

This new ComposeCreatePreviewActionK2 does not have a memory property to
avoid the memory leak. To implement it, we do not rely on
PsiBasedModCommandAction, and we implement it from the raw
ModCommandAction.

Bug: 353354965
Test: LastInIdeaTestSuite.checkForLeaks, ComposeCreatePreviewActionTest,
      ComposeSurroundWithWidgetActionTest
Change-Id: I8103829bf5bb5d7a9247e12e87e2e6205e382598