Fix OpenPluginBuildFileQuickFix when project is not loaded
OpenPluginBuildFileQuickFix could not find AndroidPluginInfo when
project first opened and not loaded properly. The reason is that it
needs to find a module but at this stage, when first sync fails,
there is no module containing build files.

However, this quickfix does not need module info anyway, just the build
file. This change adjusts code to only look for the build file.

Fixes: 343155033
Test: N/A
Change-Id: I2e7c47766da3403bbe7baa42d26b323339f3fafe