Wait for the update action to be available
This fixes the UpdateTest with the 2024.2 platform.

For some reason, the server responds an update request is performed OK
despite it being ignored by the ActionManager since the server isn't
actually ready to receive the request (component is unavailable). As
a workaround, we can wait a bit longer for the component to be ready
before trying to invoke it. Created a bug to follow up with a proper fix.

Bug: 353598800
Test: existing
Change-Id: I4cadd23cd223596bc6a5c68dc87f32d528e611ac