Avoid holding render task lock for longer than necessary
The render task lock in LayoutlibSceneManager should only guard
reads and writes of the field, but it's not needed for it to guard
the execution of RenderTask methods, because these methods are
asynchronous, so they simply return a future and the actual
execution is not being guarded.

However, this is fine, and it's responsibility of the RenderTask
to handle the synchronization needed when its rendering related
methods are executed. And the RenderTask already contains a
synchronized section under runAsyncRenderAction, which is a
private method that all render, inflate, layout and other similar
executions end up passing through.

Bug: 335424569
Test: existing should pass
Change-Id: I0e74dadd19fe8009a5ab65093bf0a415034b027c