Add infrastructure for overriding flags with Mendel experiments
(referring to the previous commits in this chain):

To get tests to work, the android-common tests had to be extracted
to a new module, and a dependency on intellij.android.testFramework
has been added. Without these changes, the new extension point in
server-flags.xml doesn't get correctly exposed during tests.

I think that this is the case because the Android plugin is
only loaded if all of its dependencies are available, and the transitive
closure of e.g. intellij.android.testFramework covers all of the
necessary dependencies.

Bug: 342183628
Test: MendelOverridesTest
Change-Id: I522f3d4d7657167c689acb2608ac8bd780eeddb2