Report sync failure for Missing Android Plugin
Add issue reporting to MissingAndroidPluginIssueChecker.
However, it only checks for missing dependency which is only generated
in case of old plugin application api.
Missing plugin id in plugins dsl results in UnknownPluginException.
Check and report it in SyncFailureUsageReporter directly as we are not
generating any issue for this case, default parser renders the error.

Bug: 354721335
Test: MissingAndroidPluginFailureTest
Change-Id: I22324e08b626ef911bd769552e4eb2adf88ed921