Move Gemini action group to ml-api
Q: "Why is there a diff in ml-api.build.gradle.kts?"

A: To make sure that the action group defined in aiplugin-api.xml
is correctly exposed in tests, each module needs to either depend on
the Android plugin or on ml-api-tests. Directly including
aiplugin-api.xml from the test plugin.xml files does not work - the
file apparently needs to be included via a dependency on a "real" plugin
even if that 'real plugin' is itself a test plugin. I've given up
trying to understand this further.

In any case, to make this work from the Gradle build, ml-api needs to
expose a test jar.

Bug: 343922465
Test: none
Change-Id: I759a17dff1cb101fcb60c3f96067d5a362694dfa