Use PSI to determine order for translation insertions
We used to determine the translation insert order from the loaded
resources. This works fine for adding translations with existing
default string values.

If a new default string is added, the translation would always be
added to the end of the translated resource file.

With this change: use the PSI of the resource item to determine the
next string in the default resource file with a translation to
determine the insert anchor.

Fixes: 338529211
Test: StringResourceDataTest
Change-Id: Ib3ecbe9e01b5d0ba256f78df824eb81296d992f2