Fix inline icon loading issue in profiler
Context: In the New UI, there are no icon versions for the inline
variants. We never made them since they were not needed in the
New UI. Thus, when accessing the WarningInline and InfoInline
Icons, a pink square showed up, reflecting the aforementioned
lack of resource.

This change removes the usage of the *Inline icons and replaces
them with their non-inline counterparts.

Fixes: 358551554
Test: manual
Change-Id: I2ee6cf542f5742d2c3ef94c85bf1f551186e1ccf