Define and implement getAndroidFacets in the project system
Rather than have the project-system-sensitive logic hidden in the
implicit behaviour of `isHolderModule`, make it explicit.  Define a
default method that simply collects all the android facets in the
project (as before, for all project systems that aren't the Gradle
one); implement the filtering in the GradleProjectSystem directly.

However, we need to be careful not to use getAndroidFacets in code
that is potentially used before the project system is set up:
otherwise we will trigger project system determination, which might be
in a situation before things are ready for it (for example in tests,
before files are copied to the filesystem).  Find and adjust such
cases.

Bug: 147586218
Test: existing
Change-Id: Icd160849469493f23a860ed5a3a8842602b48c53