Move queueing logic out of LayoutlibSceneManager
Before this change, LayoutlibSceneManager contained a custom
queueing logic equivalent to what kotlin calls a conflated
Channel, so this change replace such logic with the Channel.

Also, LayoutlibSceneManager contained a custom logic to notify
all the corresponding futures associated to a render execution
when such execution finishes. In this change that logic is being
replaced with the usage of a shared flow that every future uses
to wait for a request not older than its corresponding request
to be processed.

Note that this two logics are being modified together because they
are closely related to each other.

Bug: 335424569
Test: LayoutlibSceneRendererTest
Change-Id: I987d94971c95a343e9a3625743790ae823d011e7