Make sure a delayed selection change is eventually fired
When we expand/collapse a node, the code will also select the node
being expanded/collapsed. However the selection change was not sent
to the Layout Inspector. Make sure that we eventually send an
actual selection change.

Also: When we collapse and expand a node we should not cause a
goto declaration action. Avoid that by only generating a double
click event when clicking on the actual item (not the expander).

Fixes: 360803450
Test: Existing tests adapted to cover
Change-Id: I4a6a1417e20d23f2d18cffbf11da3751082633dc