Resolve view binding layouts for androidTest
Generated view binding classes for views defined within androidTest are
under the test package name, not the main module's package name. Up
until now, we weren't generated the light classes for these correctly.

This is partially because when the bug was noticed, resources for
androidTest were entirely broken. ag/24557715 and other changes fixed
the resource system for androidTest, so now it's much more
straightforward to simply define the correct package names for
androidTest binding classes.

Fixes: 150221074
Test: DataBindingInspectionVerificationTest.verifyNoFilesShowInspectionErrors
Change-Id: Ibf96626efa582c5bd81536181c0e4ca2bfb0c122