Remove requestLayoutAndRenderAsync
This method is not necessary as it ended up calling RenderAsync
in most cases. And in the few cases were it called layout instead,
this call of layout can be replaced by a render.

Bug: 335424569
Test: existing should pass and manually verified that the clients
of this method still work well (motion play/pause, layout editor
live updates, layout editor scroll view interactions, color blind
modes in layout editor and previews)

Change-Id: I0c4d1da88be4e9e0e31b3826fb707fa9d0ab6f00