Handle changing list of devices in filters
The current set filters build a complete list of their values at
initialization time. However, we want to allow the list of devices to
change. This means there may be values that were not present at
initialization time. When devices with new values arrive, we should
add the new values to the set filter, defaulting to on.

Bug: 334967743
Test: DeviceTableTest
Change-Id: I4a47ab666aaec2c4ab899a22ad9d46e89a1334c3