Work around leak in XDebuggerManager
XDebuggerManager is a preloaded service, meaning that IntelliJ will
load it in a background thread upon project open. Unfortunately, if a
test finishes while XDebuggerManager is in the middle of
initialization, then it may leave some disposables dangling in the
disposer tree.

As a workaround, we can wait for XDebuggerManager initialization
explicitly. Note, it is currently unknown why
AndroidImportProjectActionTest seems to be the only affected test.

b/362785921 tracks future work, potentially fixing the issue upstream.

Bug: 362785921
Fixes: 362335169
Fixes: 356483696
Test: intellij.android.core.actions_tests (500x)
Change-Id: I216473cfa2cb55ed20a4c26b0a72589c9cc31324