Compute GradleFiles changed status on fileOpened
If some non-file-editor changes the contents of a gradle file since
the last sync / computation of file hashes, then a user opening the
editor on their build file will be surprised by the lack of a banner.

Checking the status on fileOpened is not a complete solution, because
there will still be cases where we do not have hashes (for instance,
when opening a project from cache but with changed gradle files) and
so will fail to detect the need to sync.  This does, however, solve
the problem of in-IDE changes (for example, manual changes from the
commit diff window).

Fixes: 345518607
Test: manual
Change-Id: Ia37812b57a90d5960e74afdee18fc98f1445d8ee