Add enable code context dialog flow
This change adds support for the showing of the enable code context
button in the bottom toolbar of insights. There are a couple of other
changes bundled:

- New event called EnhanceInsight.
- Updated code context tracking to take into account whether or not the
current insight is generated with additional code context. With the new
user experience, it is not always the case that a user assigned to
experiment X will always see insights generated with X.
- In addition, users assigned to the control experiment will not ever
see the enable context button.
- UI and controller changes to facilitate the showing and the hiding of
the enable context button.

Known issues:
- Somtimes, the button appears briefly before disappearing.

Test: new tests.
Bug: 364939650
Change-Id: I0658b8bbf668e5a5559ca9058fa9edd6eede9355