Only enable or disable sensors when outside an exercise
... when applying changes.

This is to prevent the scenario: enable/disable a sensor before the
exercise starts without applying. Once the exercise starts the change is
still pending. Click apply.

Technically, this is fine as the call succeeds and the content provider
is updated. However, it can be confusing because:
1. we don't allow enabling/disabling sensors once in an exercise
2. if the sensor is disabled or enabled, it will only take effect once
   the next exercise starts, not the current one

There is another issue which will be corrected in the next CL: the
pending change adds an asterisk next to the capability due to a pending
change, even though the checkbox is disabled.

Bug: 338446522
Test: //tools/adt/idea/wear-whs:intellij.android.wear-whs_tests
Change-Id: I1799b12a0d63b59fd8f07fa085b689f72f40c867