Parameterize DeviceSource by profile type
Since we are no longer mixing DeviceProfiles of various types in a
dialog, it becomes useful to parameterize the DeviceTable by the type of
profiles it contains. This will allow us to more naturally access the
type-specific fields of the profiles as we diverge the dialogs further.
It already allows us to remove a few casts.

Delete the no-longer-useful "source" field.

Bug: 335264656
Test: Updated
Change-Id: Iea1611f73da2b8c388add6b9e834ce1240a236e0