Remove obsolete deprecate warning in SurfaceLayoutManager
SceneViewLayoutManager does not exist as a class nor an interface, and
an instance of NlDesignSurfacePositionableContentLayoutManager defined
in NlDesignSurface is the only thing I could find that could be the
meaning of this deprecation message. It seems we shouldn't show the
deprecation at all, since SurfaceLayoutManager is largely used and there
is no apparent reason for its deprecation.

Bug: N/A
Test: N/A
Change-Id: Ic6f595700957836dfa8886b07f7002ae968eefb5