[android] use LF line separators instead of CRLF
Currently, line separators in text files are specified via git config. This isn't convenient because it requires additional configuration on each location where the source is checked out. It would be more convenient to specify line separators via .gitattributes file, but for that we need to ensure that line separators in files are consistent with .gitattributes. Most of the files already use LF separators, this commit converts line separator to LF for several files which use CRLF.

This commit is an accumulation of the following cherry-picks:
  - commit 46c8d5a22c19fb2d9592a6f6b92cc3aef33106b7
  - commit 2111225cc7b0b50dec51f1486c6f93a47c7c7194
  - commit cf5b308fcda3b027cd9459030e23254a6a3d640e
  - commit ebab3f37d68a3bc367db6420aa4c8a1f89a94b25
  - commit e237656c72842f5e1f6baffb95da7eedbdd23fca

GitOrigin-RevId: 7ad36716c8f1baa07699c20439cba051170589d5
Change-Id: I6740cfc2f30a6c7d2b3da04bbe7ae9dbf3d8bc9f