Use higher-level GraphicsMode instead of GpuMode
The UI always displays "auto", "hardware", or "software". The "software"
setting gets translated into one of two GpuModes depending on the system
image. Since we don't know which system image will be chosen until the
end, use a separate enum that reflects the UI setting, and translate it
into the appropriate GpuMode when we update the device.

Fixes: 364350803
Test: existing
Change-Id: I630fb365fddff7546e00d48131f269ac73fa3400