Fix PreviewCanvasTracker not uploading metrics
We were not reporting the collected metrics, because we were passing
null as the tracker key. The intent was to share the key between all
Compose Previews, but if null is passed as a key, we'll default to the
no-op tracker, which is supposed to be used by tests only.

This change fixes the issue by passing the current DesignSurface as the
key. The surface is unused for this tracker specifically, but this way
we ensure the metrics are tracked and that the tracker is disposed
properly now that's tied to a surface.

Bug: N/A
Test: PreviewCanvasTrackerTest and SwitchSurfaceLayoutManagerActionTest.
 Also manually checked that metrics are tracked.
Change-Id: Ie0ea1df927b6bbd04ca4a6e9ae314c3ec9ece76a