Move Gradle-using test to a target that has Gradle available
The only use of Gradle in this test is to open a project that fails
sync, so in some sense running it in a test target without access to
Gradle at all is correct.  However, it does confuse the test log and
will be brittle to subsequent modifications, so it's probably better
to put it in a test target that does have Gradle available.

Bug: none filed
Test: existing
Change-Id: I2d00e59a2f21d9a7240bb0a82a65ee3465527d30