Enquire directly about the production state of the module
This doesn't (currently) change any behaviour: the implementation of
isProductionAndroidModule() is (for modules with AndroidFacets) the
same as asking whether it is the "main module" in all project systems.
A difference might come in in a future change where we would no longer
consider a test-only (Gradle) project to have a "main" module.

Bug: 147586218
Test: existing
Change-Id: I5c808f6f935ef5558e0e04d6fe802e3280a9293d