Ensure all jars are attributed to the right target.
In the aspect, instead of assuming that all jars provided by
`JavaInfo.compile_jars` and `JavaInfo.transitive_compile_time_jars` come
from the current target, use the `file.owner` API to attribute them to
the target that actually produced them.

To achieve this, we don't add these file to the `target_to_artifacts`
map as we encouter them, but add them after we've collect all
dependencies instead, using the depset of all jars that we build anyway.

This ensures that the downstream IDE code correctly attributes all jar
files to the actual target that built them, which in turns allows the
jar to source resolution implemented inside `ArtifactIndex` to work
properly. That is used by `QuerySyncNavigationPolicy`, so fixing it
ensures "go to definition" actions works reliably for external
dependencies.

To test this change, we also add support for unit testing of the aspect
code.

Add support creating rules to create a `.java-info.txt` file to
simulate build deps for a set of targets, with a custom project scope
(i.e. set of include directories). Add support to build_dependencies.bzl
for this, as required to work around the restrictions that bazel places
on aspect parameters when the aspect is invoked from a build rule.

Also add a unit test to assert some things about the aspect output that
we want to be true:
- each jar file is listed under a single target
- a target than exports other targets should not contain those exported
  jars

Bug: 353651427
Test: Unit test included, manually verified.
Change-Id: I2caea196373a9d9fed47164c0f9ef97cfc885cb9