Show actual RAM size in editor
When creating an AVD based on a device profile, we will often default to
a RAM size that is smaller than what the device profile contains.

However, when editing the device profile, the editor currently
substitutes this default value for the actual device value. This is
confusing; we should just show the actual value (even though it doesn't
get used for much).

Fixes: 181336801
Test: AvdDeviceDataTest
Change-Id: I852d8c578ec16189f87aa155579c11a6f111dd6a