Fix bug with expanded storage validation
The LaunchedEffect was triggering on the initial state of the Existing
Image text, even though the radio button for Existing Image was not
selected.

Consolidate the UI state into the domain objects in a snapshot flow and
publish those, instead of triggering updates directly from each UI
element.

Switch from Path to String since we're not actually preserving the
FileSystem anyway and it complicates things.

Fixes: 367396730
Test: AdditionalSettingsPanelTest
Change-Id: I5a718af59c934f805f00d9a951bb9cca6a427367