Adapt RunConfigurationProducers to the new conventions
Make sure that if they're finding modules from a context they perform
conversion to a holder module.

Since it is now an unexpected but not fatal condition that the
testPackageName not be available for a run configuration, convert a
logger error (IDE / test crash) into a warning.

Adjust a test (of a library module with no androidTest artifact in
release mode) to make sure that `validate` is called with the right
access permissions, and also for the change in error message.

Bug: 147586218
Test: existing test adjusted
Change-Id: Ie77ddf9e2ad7d484a59c95c013187b61f2d48dc9