Allow sub-modules for creating resource in XML
Change ag/26053286 removed the ability of the "Create new resource"
dialog to specify that the resource be created in a sub-module in order
to fix b/318743105. The fix was meant for Java and Kotlin, since in
those circumstances creating the resource in a sub-module resulted in
incorrect, non-compiling code.

But there's no reason we can't allow this for XML files, since the
resource reference doesn't have to change in those files. This change
adds back the sub-module selection for XML resource files.

Fixes: 365591190
Test: Manually verified
Change-Id: Ib9b7631fc3f09a2ebd5685e5e8517528954651c8